pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "opentrans"
dynamic = ["version"]
authors = [
  { name="Hydrangea" },
]
description = "A directory-preserving AI translator using Ollama."
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
keywords = ["translation", "ollama", "llm", "ai", "automation"]
dependencies = [
  "ollama>=0.1.0",
  "tqdm>=4.66.0",
  "click",
  "pydantic",
  "pydantic-settings",
  "pyyaml",
  "mkdocs-static-i18n>=1.3.0",
  "tree-sitter<0.22",
  "tree-sitter-markdown>=0.5.1",
  "tree-sitter-languages>=1.10.2",
]
classifiers = [
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Programming Language :: Python :: 3.14",
    "Operating System :: OS Independent",
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Topic :: Text Processing :: Linguistic",
    "License :: OSI Approved :: MIT License",
]

[project.urls]
Homepage = "https://github.com/RainStorm108/opentrans"
Documentation = "https://rainstorm108.github.io/OpenTrans/"
Issues = "https://github.com/RainStorm108/opentrans/issues"

[project.scripts]
opentrans = "opentrans.main:main"

[tool.hatch.version]
path = "src/opentrans/__init__.py"

[tool.hatch.build.targets.sdist]
exclude = [
  "/test",
  "/.venv",
  "/.vscode"
]

[tool.hatch.build.targets.wheel]
packages = ["src/opentrans"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]

# Enable live logging to the terminal
log_cli = true
log_cli_level = "INFO"
log_cli_format = "%(asctime)s [%(levelname)8s] %(message)s (%(filename)s:%(lineno)s)"
log_cli_date_format = "%Y-%m-%d %H:%M:%S"
